// app/src/api/generated/models/channel.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Tube Archivist API
 * API documentation for Tube Archivist backend.
 * OpenAPI spec version: v0.5.10
 */
import type { ChannelOverwrite } from './channelOverwrite';
import type { VidTypeFilterEnum } from './vidTypeFilterEnum';

/**
 * serialize channel
 */
export interface Channel {
  channel_id: string;
  channel_active: boolean;
  /** @nullable */
  channel_banner_url?: string | null;
  /** @nullable */
  channel_thumb_url?: string | null;
  /** @nullable */
  channel_tvart_url?: string | null;
  /** @nullable */
  channel_description?: string | null;
  channel_last_refresh: string;
  channel_name: string;
  channel_overwrites?: ChannelOverwrite;
  channel_subs: number;
  channel_subscribed: boolean;
  channel_tags?: string[];
  channel_tabs: VidTypeFilterEnum[];
  _index?: string;
  _score?: number;
}