// app/src/api/generated/models/playlist.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Tube Archivist API
 * API documentation for Tube Archivist backend.
 * OpenAPI spec version: v0.5.10
 */
import type { PlaylistEntry } from './playlistEntry';
import type { PlaylistSortOrderEnum } from './playlistSortOrderEnum';
import type { PlaylistTypeEnum } from './playlistTypeEnum';

/**
 * serialize playlist
 */
export interface Playlist {
  playlist_active: boolean;
  playlist_channel: string;
  playlist_channel_id: string;
  /** @nullable */
  playlist_description?: string | null;
  playlist_entries: PlaylistEntry[];
  playlist_id: string;
  playlist_last_refresh: string;
  playlist_name: string;
  playlist_subscribed: boolean;
  playlist_sort_order: PlaylistSortOrderEnum;
  playlist_thumbnail: string;
  playlist_type: PlaylistTypeEnum;
  _index?: string;
  _score?: number;
}