// app/src/api/generated/models/playlistEntry.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Tube Archivist API
 * API documentation for Tube Archivist backend.
 * OpenAPI spec version: v0.5.10
 */

/**
 * serialize single playlist entry
 */
export interface PlaylistEntry {
  youtube_id: string;
  title: string;
  /** @nullable */
  uploader: string | null;
  idx: number;
  downloaded: boolean;
}