// app/src/api/generated/models/video.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Tube Archivist API
 * API documentation for Tube Archivist backend.
 * OpenAPI spec version: v0.5.10
 */
import type { Channel } from './channel';
import type { Player } from './player';
import type { SponsorBlock } from './sponsorBlock';
import type { Stats } from './stats';
import type { StreamItem } from './streamItem';
import type { SubtitleItem } from './subtitleItem';
import type { VidTypeFilterEnum } from './vidTypeFilterEnum';

/**
 * serialize video item
 */
export interface Video {
  active: boolean;
  category: string[];
  channel?: Channel;
  /** @nullable */
  comment_count?: number | null;
  date_downloaded: number;
  /** @nullable */
  description?: string | null;
  media_size: number;
  media_url: string;
  player: Player;
  playlist?: string[];
  published: string;
  sponsorblock?: SponsorBlock | null;
  stats: Stats;
  streams: StreamItem[];
  subtitles?: SubtitleItem[];
  tags: string[];
  title: string;
  vid_last_refresh: string;
  vid_thumb_url: string;
  vid_type: VidTypeFilterEnum;
  youtube_id: string;
  _index?: string;
  _score?: number;
}